added stream buffer to scanner
putback() and sip() will push and pop to and from the buffer. The buffer
holds characters from the input stream. When we want to "return
something to the stream" we push it into the buffer. When we want to get
a char from the stream we first check if we have something in the buffer
before we turn to the stream.

This is because in C we can't really (for what I gather) fputc() a char
back into the stream and then expect to get that char back when we
fgetc() again. The buffer is to provide this mechanism.