fixed cl input file handling
the input files specified at the commandline. the list parse_cl_opts()
generated was corrupt.

still, when a file can't be opened parse_cl_opts() returns an error
code, and main() then aborts. should this be an error, or just a warning
and we proceed with processing any other files?

when a file can't be opened parse_cl_opts() now simply doesn't create a
link for it in opts->input_files, so it simply doesn't exist in the list
of files to process.