initial import of plaintext generator
the target the user specifies on the commandline sets the gen function
pointer inside the opts (struct conf) to the proper generator function.
this function, given an output stream to talk to and a syntax tree, will
spit out the "code" generated for the relevant target.

i start as an exercise with the plaintext target, as it's the easiest.
all i do is convert the headings to all-caps. it outputs the rest of the
text as is.

notice i got rid of the target enum in struct conf. i find it more
convenient to simply have a fn ptr in the struct and have the option
parsing function hook it up immeidately with the appropriate generation
function.