renamed sip() to scan(), scan() to forced_scan()
in parsers it is more common to call the scanning function scan().
forced_scan(), the gigantic function that actually does the dirty
scanning work, is now called forced_scan() because it will actually go
through the scanning process when called. This is as opposed to the new
scan() which may or may not perform a scan depending on whether there's
a cached token waiting in the token buffer, lstate->token_buf.