makefile: don't link if all is done
this "fix" slows down the build by about 1.8% (7ms atm) because before
we compile each compilation unit we verify the parent directory exists.
Previously we checked only once, before starting compiling, but that
implied that the directory structure creation rule had to be .PHONY,
which in turn implied the TARGET rule had to be done. That's what caused
the "PHONY" linking.