added scripts for autogenerating test foundations
with these scripts i hope that now for writing tests all i need to do is
just write the tests themselves. code that is within TEST(test_name) { }
blocks. then i can run the scripts to generate the code that plugs the
test code into the testing process: declare the tests and create a test
suite in a header file, add the suite to the other suites in the main.c
file, and write the main() function that runs the tests.