gen_test_main.sh: allow setting output prefix
Run the script with the -o option to set the prefix for regression
output. For instance, run

./gen_test_main.sh -o test/reg/

to set the test/reg as the output directory where the output of tests is
stored. If a test suite named test_string has a test with a line such as
printf("Hi") then CU will write Hi into the file
test/reg/tmp_test_string.out. Similarly CU will create a file with the
extension .err for output to stderr.

The CU framework supports specifying any prefix, but the script will
always append a trailing slash if you didn't specify one yourself, so
that the prefix is always a directory.