fix: "overflowed indent" bug
if a heading was indented more than necessary (say, start a document
with two spaces followed with a heading text) then only the necessary
indents were parsed (one indent in the previous example) and the other
indents were then unexpected. instead, the scanner should notice when
the indent is larger than necessary and drop the extra whitespace indent
so that it doesn't scan this redundant whitespace as indent in the next
scan call.